:truck: separate folders based on module categories
